Added script time limit - requires a patch to Lua, but the one one LuaWiki is broken. Increased SoundManager accuracy to 1/43rd of a second Added counters Detemplated Logger Vector stream operators so they now work. Removed legacy stuff at last
Added 'Loss of thrust factor' to core, slows you down when you take damage
Added Geometry Wars 2.luaChallenge
Improved I Lurk.LuaChallenge
Added TinyXML fix to Scorekeeper - this needs a customised TinyXml version :(
Upgraded to Lua5.13, mostly to apply the thread abort patch

+++ OGLTest_lib/KeyboardAI.h
@@ -16,6 +16,6 @@ public:
 	virtual void SpecifyTarget(Core* _target){}
 	virtual bool IsHuman(){return true;}
 
-	static int GetInstanceCount(){return player_ids_.size();}
+	static int GetInstanceCount(){return static_cast<int>(player_ids_.size());}
 	static std::vector<int> GetPlayerIDs(){return player_ids_;}
 };
